Rearrange e2e Dockerfile for better caching
When building an image, Docker caches after each line in the Dockerfile.
If there are no changes to this line (or above), then it can build the
image from a component in the cache. This caching is particularly
helpful with expensive install operations.

Installing `ansible` is extremely expensive. Because we specified
Ansible as a dependency through `setup.py`, we had to reinstall it
everytime we changed the code (because changing the code invalidated
the Docker image cache). However, by installing ansible (and all of
`requirements.txt`) at the start of the Dockerfile, we cache these
dependencies. We unfortunately cannot move `setup.py` into this "early"
run, as it depends on the code in `sheepdog`.