Refactor `Config` class as singleton
The previous role of the `Config` class was not clearly defined. It was
responsible for managing configuration, but only some classes could
access it. This lack of accessibility meant methods often had a number
of parameters, and we often had to pass parameters through multiple
methods before they arose at the method which actually used them.

Instead, we implement the singleton pattern, so we initialize the
configuration once with default/runtime values, and then allow multiple
consumers to read from it. Additionally, we allow class constructors to
either accept instances of `Config` as an argument or to get the current
singleton. This grants us flexibility during testing.