Support bootstrap and cron options for `kennel run`
Fix #8

sheepdog provides control over the pups included in each runs. The pups
with the bootstrap tag will only be run when the `--run-mode` is
explicitly set to `bootstrap`. The pups with the cron tag will have a
cron job set to run them at a regularly scheduled interval. This
granularity allows you to use the same kennel to perform one-time
machine provisioning operations, in addition to tasks you want to run
automatically.

This commit adds the `--run-mode` option to `kennel run` and updates the
current code/tests accordingly.

The next commit will include an integration test of the `bootstrap/cron`
functionality, in addition to adding the cron job creation to
`pup-base`.