Add support for installing python and ansible deps
`pups` can now specify python and ansible dependencies using
`requirements.{txt,yml}`. We will run the relevant `install` command on
these files.

We made the decision to not try and implement the virtualenv for the
user when running install. At least initially, it will be the
responsibility of the user to ensure they've activated the proper
virtualenv when running sheepdog. We do provide a mechanism for
activating a virtualenv during the cron run.