Sync kennel with git before cron run
Fix #20

If a user manages their kennel through Git, and applies the same kennel
to multiple machines, the regularly scheduled cron run should pick up
the most recent changes. Set up the cron job to check if the git manages
the kennel, and if yes, run a `git pull origin master` and a `sheepdog
install` before the `sheepdog run` command.