Refactor cli to use `click` group
Previously, we had only one `click` action, which we tried
to use to execute all CLI actions. This usage was hard to extend, and
not an idiomatic use of `click`.

Instead, we define a `cli` `click` group, and then a number of different
commands on the group. We do this in preparation for adding the `create`
cli command.