Succeed integration test as soon as cron job runs
To test that sheepdog's cron functionality works as desired, a couple
of our integration tests must wait for a cron run to occur before they
can succeed. Previously, we waited a set amount of times, 2 minutes, and
then checked if the job run. With this implementation, we waited the
same amount of time, even if the cron job executed earlier.

Update the `assert_e2e_state.sh` method so it immediately starts
checking if the cron job is complete. It succeeds if it ran
successfully, and times out with an error if it doesn't succeed after a
set amount of time.