Convert sheepdoge to using pipenv
Convert sheepdoge to using the pipenv project. This allows me to
create a Pipfile.lock, ensuring replication of dependencies.
We removed `requirements.txt`, although kept `setup.py` as its needed
for `pip install`.

We updated the `Makefile` accordingly as well as adding a `pipenv check`
to the `.travis.yml` so we consistently check that our dependencies
are secure and valid.