Fix bug where we required vault passwd file
Previously, we assumed a vault password file existed. The user had to
create that file (even if we didn't actually use if for decrypting
anything), or else ansible would fail.

Update so we only try and include the --vault-password-file` option in
the ansible run if the user specifies the vault password file.