Use python3 by default for sheepdoge dev
By default, I want to use python3 for sheepdoge development. However, I
want to remain compatible with python2.7 (given the ansible default
seems to still be 2.7). To accomplish this I made the following changes:

- Specify `pipenv` should use the default environment. For me locally,
this is python3. On travis, this will be both 2.7 and 3.6.
- Fix any issues associated with running python3 code.
- Update `setup.py` so its clear we can use python3.
- Update all integration tests to use python3, and add a new integration
test which uses python2.

As a small side benefit, I switched all my base Docker images for
testing to the standard ubuntu:16.04, meaning I no longer have to do all
of the weird PYTHONPATH manipulation in the `install_cron-Debian.yml`
task.