Rename vars to reflect new var files
We name variables in the kennel/pup relative to the var file they go in.
Since we are changing the names of the var files to `main` and
`encrypted`, we want to update the var names accordingly in all of our
integration tests. For each pup/kennel, I'll do a separate commit to
update the variable names.